split test file into more specific files
