+++ test/global.js
@@ -1,0 +1,7 @@
+require('lie/polyfill');
+
+var chai = require("chai");
+var chaiAsPromised = require("chai-as-promised");
+chai.use(chaiAsPromised);
+
+global.expect = chai.expect;
